# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RESMA Programmatic and Agency Report API",
    "version": "0.2.0",
    "summary": "Stable aggregate-only machine surface for analysis, agency report delivery, and AI workflows.",
    "description": "RESMA exposes aggregate-only observatory intelligence for AI clients, agency operators, and premium programmatic consumers. The API is a delivery channel for package-gated aggregate intelligence, not a raw contributor-data surface. Contributor-level feeds, per-user traces, and raw audience exports are intentionally excluded."
  },
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development server"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "Programmatic API key. `Authorization: Bearer <api-key>` is also accepted."
      }
    },
...
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "const": false
          },
          "error": {
            "type": "string"
          },
          "details": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": true
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
...
      "SuccessEnvelope": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          },
          "meta": {
            "type": "object",
            "additionalProperties": true
          },
          "format": {
            "type": "string",
            "enum": [
              "llm",
              "markdown",
              "client-report"
            ]
          },
          "llm": {
            "type": "object",
            "properties": {
              "kind": {
                "type": "string"
              },
              "title": {
                "type": "string"
              },
              "bullets": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "markdown": {
                "type": "string"
              },
              "followUpQuestions": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "caveats": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "additionalProperties": false
          },
          "export": {
            "type": "object",
            "properties": {
              "format": {
                "type": "string"
              },
              "content": {
                "oneOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "object",
                    "additionalProperties": true
                  }
                ]
              }
            },
            "additionalProperties": true
          }
        },
        "required": [
          "success",
          "data"
        ],
        "additionalProperties": true
      }
    }
  },
  "paths": {
...
    "/api-keys": {
      "get": {
        "tags": [
          "API Keys"
        ],
        "summary": "List API keys for the authenticated user",
        "operationId": "listApiKeys",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current API keys and quota snapshots",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "API Keys"
        ],
        "summary": "Create a new API key scoped to the caller package",
        "operationId": "createApiKey",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 80
                  },
                  "scopes": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "dailyQuota": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "monthlyQuota": {
                    "type": "integer",
                    "minimum": 1
                  },
                  "expiresAt": {
                    "type": "string",
                    "format": "date-time"
                  }
                },
                "required": [
                  "name"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "API key created. Secret is returned once.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api-keys/{apiKeyId}": {
      "delete": {
        "tags": [
          "API Keys"
        ],
        "summary": "Revoke an API key",
        "operationId": "revokeApiKey",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "apiKeyId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Key revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Key not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  }
}